APIEX3/requirements.txt
python-dotenv
beaupy
requests